The man who was Trump presidential decree of Prime Minister Abe.
Chinese economy market make money  where the real estate bubble bursts. and the land can not be purchased, there is stability in the Japanese economy that would require the G5 in the UK because so USA is cautious on the military industry. It was a reverse policy that made a big turn  the north of President Trump. If there is negotiations with the North again, it will depend on the revision of Japan. Prime Minister Abe's LDP should drop his seat. Distrust of the party. It will be the social distrust  the rendezvous with  Prime Minister Abe's President Trump. If there is a revision Japanese constitution of Japan, it will be an election for the abolition of the emperor. Compliance with the peace constitution should be considered national interest.